+++ sdk/contentsafety/ai-content-safety-rest/samples/v1/javascript/sampleManageBlocklist.js
@@ -5,42 +5,29 @@
  * @summary Demonstrates how to manage blocklist.
  */
 
-const ContentSafetyClient = require("@azure-rest/ai-content-safety").default,
-  { isUnexpected } = require("@azure-rest/ai-content-safety");
+const { BlocklistClient, ContentSafetyClient } = require("@azure-rest/ai-content-safety");
 const { AzureKeyCredential } = require("@azure/core-auth");
-
-// Load the .env file if it exists
-require("dotenv").config();
+require("dotenv/config");
 
 const endpoint = process.env["CONTENT_SAFETY_ENDPOINT"] || "<endpoint>";
 const key = process.env["CONTENT_SAFETY_API_KEY"] || "<key>";
 
 const credential = new AzureKeyCredential(key);
-const client = ContentSafetyClient(endpoint, credential);
+const contentSafetyClient = new ContentSafetyClient(endpoint, credential);
+const blocklistClient = new BlocklistClient(endpoint, credential);
 
 async function createOrUpdateTextBlocklist() {
   const blocklistName = "TestBlocklist";
   const blocklistDescription = "Test blocklist management.";
-  const createOrUpdateTextBlocklistParameters = {
-    contentType: "application/merge-patch+json",
-    body: {
-      description: blocklistDescription,
-    },
-  };
-
-  const result = await client
-    .path("/text/blocklists/{blocklistName}", blocklistName)
-    .patch(createOrUpdateTextBlocklistParameters);
-
-  if (isUnexpected(result)) {
-    throw result;
-  }
+  const result = await blocklistClient.createOrUpdateTextBlocklist(blocklistName, {
+    description: blocklistDescription,
+  });
 
   console.log(
     "Blocklist created or updated: Name",
-    result.body.blocklistName,
+    result.blocklistName,
     ", Description: ",
-    result.body.description,
+    result.description,
   );
 }
 
@@ -48,32 +35,22 @@ async function addBlockItems() {
   const blocklistName = "TestBlocklist";
   const blockItemText1 = "sample";
   const blockItemText2 = "text";
-  const addOrUpdateBlocklistItemsParameters = {
-    body: {
-      blocklistItems: [
-        {
-          description: "Test block item 1",
-          text: blockItemText1,
-        },
-        {
-          description: "Test block item 2",
-          text: blockItemText2,
-        },
-      ],
-    },
-  };
-
-  const result = await client
-    .path("/text/blocklists/{blocklistName}:addOrUpdateBlocklistItems", blocklistName)
-    .post(addOrUpdateBlocklistItemsParameters);
-
-  if (isUnexpected(result)) {
-    throw result;
-  }
+  const result = await blocklistClient.addOrUpdateBlocklistItems(blocklistName, {
+    blocklistItems: [
+      {
+        description: "Test block item 1",
+        text: blockItemText1,
+      },
+      {
+        description: "Test block item 2",
+        text: blockItemText2,
+      },
+    ],
+  });
 
   console.log("Block items added: ");
-  if (result.body.blocklistItems) {
-    for (const blockItem of result.body.blocklistItems) {
+  if (result.blocklistItems) {
+    for (const blockItem of result.blocklistItems) {
       console.log(
         "BlockItemId: ",
         blockItem.blocklistItemId,
@@ -89,23 +66,15 @@ async function addBlockItems() {
 async function analyzeTextWithBlocklists() {
   const blocklistName = "TestBlocklist";
   const inputText = "This is a sample to test text with blocklist.";
-  const analyzeTextParameters = {
-    body: {
-      text: inputText,
-      blocklistNames: [blocklistName],
-      haltOnBlocklistHit: false,
-    },
-  };
-
-  const result = await client.path("/text:analyze").post(analyzeTextParameters);
-
-  if (isUnexpected(result)) {
-    throw result;
-  }
+  const result = await contentSafetyClient.analyzeText({
+    text: inputText,
+    blocklistNames: [blocklistName],
+    haltOnBlocklistHit: false,
+  });
 
   console.log("Blocklist match results: ");
-  if (result.body.blocklistsMatch) {
-    for (const blocklistMatchResult of result.body.blocklistsMatch) {
+  if (result.blocklistsMatch) {
+    for (const blocklistMatchResult of result.blocklistsMatch) {
       console.log(
         "BlocklistName: ",
         blocklistMatchResult.blocklistName,
@@ -119,141 +88,90 @@ async function analyzeTextWithBlocklists() {
 }
 
 async function listTextBlocklists() {
-  const result = await client.path("/text/blocklists").get();
-
-  if (isUnexpected(result)) {
-    throw result;
-  }
-
   console.log("List blocklists: ");
-  if (result.body.value) {
-    for (const blocklist of result.body.value) {
-      console.log(
-        "BlocklistName: ",
-        blocklist.blocklistName,
-        ", Description: ",
-        blocklist.description,
-      );
-    }
+  for await (const blocklist of blocklistClient.listTextBlocklists()) {
+    console.log(
+      "BlocklistName: ",
+      blocklist.blocklistName,
+      ", Description: ",
+      blocklist.description,
+    );
   }
 }
 
 async function getTextBlocklist() {
   const blocklistName = "TestBlocklist";
 
-  const result = await client.path("/text/blocklists/{blocklistName}", blocklistName).get();
-
-  if (isUnexpected(result)) {
-    throw result;
-  }
+  const result = await blocklistClient.getTextBlocklist(blocklistName);
 
   console.log("Get blocklist: ");
-  console.log("Name: ", result.body.blocklistName, ", Description: ", result.body.description);
+  console.log("Name: ", result.blocklistName, ", Description: ", result.description);
 }
 
 async function listBlockItems() {
   const blocklistName = "TestBlocklist";
 
-  const result = await client
-    .path("/text/blocklists/{blocklistName}/blocklistItems", blocklistName)
-    .get();
-
-  if (isUnexpected(result)) {
-    throw result;
-  }
-
   console.log("List block items: ");
-  if (result.body.value) {
-    for (const blockItem of result.body.value) {
-      console.log(
-        "BlockItemId: ",
-        blockItem.blocklistItemId,
-        ", Text: ",
-        blockItem.text,
-        ", Description: ",
-        blockItem.description,
-      );
-    }
+  for await (const blockItem of blocklistClient.listTextBlocklistItems(blocklistName)) {
+    console.log(
+      "BlockItemId: ",
+      blockItem.blocklistItemId,
+      ", Text: ",
+      blockItem.text,
+      ", Description: ",
+      blockItem.description,
+    );
   }
 }
 
 async function getBlockItem() {
   const blocklistName = "TestBlocklist";
   const blockItemText = "sample";
-  const addOrUpdateBlocklistItemsParameters = {
-    body: {
-      blocklistItems: [
-        {
-          description: "Test block item 1",
-          text: blockItemText,
-        },
-      ],
-    },
-  };
-  const result = await client
-    .path("/text/blocklists/{blocklistName}:addOrUpdateBlocklistItems", blocklistName)
-    .post(addOrUpdateBlocklistItemsParameters);
-  if (isUnexpected(result) || result.body.blocklistItems === undefined) {
+  const result = await blocklistClient.addOrUpdateBlocklistItems(blocklistName, {
+    blocklistItems: [
+      {
+        description: "Test block item 1",
+        text: blockItemText,
+      },
+    ],
+  });
+  if (result.blocklistItems === undefined) {
     throw new Error("Block item not added.");
   }
-  const blockItemId = result.body.blocklistItems[0].blocklistItemId;
+  const blockItemId = result.blocklistItems[0].blocklistItemId;
 
-  const blockItem = await client
-    .path(
-      "/text/blocklists/{blocklistName}/blocklistItems/{blocklistItemId}",
-      blocklistName,
-      blockItemId,
-    )
-    .get();
-
-  if (isUnexpected(blockItem)) {
-    throw blockItem;
-  }
+  const blockItem = await blocklistClient.getTextBlocklistItem(blocklistName, blockItemId);
 
   console.log("Get blockitem: ");
   console.log(
     "BlockItemId: ",
-    blockItem.body.blocklistItemId,
+    blockItem.blocklistItemId,
     ", Text: ",
-    blockItem.body.text,
+    blockItem.text,
     ", Description: ",
-    blockItem.body.description,
+    blockItem.description,
   );
 }
 
 async function removeBlockItems() {
   const blocklistName = "TestBlocklist";
   const blockItemText = "sample";
-  const addOrUpdateBlocklistItemsParameters = {
-    body: {
-      blocklistItems: [
-        {
-          description: "Test block item 1",
-          text: blockItemText,
-        },
-      ],
-    },
-  };
-  const result = await client
-    .path("/text/blocklists/{blocklistName}:addOrUpdateBlocklistItems", blocklistName)
-    .post(addOrUpdateBlocklistItemsParameters);
-  if (isUnexpected(result) || result.body.blocklistItems === undefined) {
+  const result = await blocklistClient.addOrUpdateBlocklistItems(blocklistName, {
+    blocklistItems: [
+      {
+        description: "Test block item 1",
+        text: blockItemText,
+      },
+    ],
+  });
+  if (result.blocklistItems === undefined) {
     throw new Error("Block item not added.");
   }
-  const blockItemId = result.body.blocklistItems[0].blocklistItemId;
+  const blockItemId = result.blocklistItems[0].blocklistItemId;
 
-  const removeBlocklistItemsParameters = {
-    body: {
-      blocklistItemIds: [blockItemId],
-    },
-  };
-  const removeBlockItem = await client
-    .path("/text/blocklists/{blocklistName}:removeBlocklistItems", blocklistName)
-    .post(removeBlocklistItemsParameters);
-
-  if (isUnexpected(removeBlockItem)) {
-    throw removeBlockItem;
-  }
+  await blocklistClient.removeBlocklistItems(blocklistName, {
+    blocklistItemIds: [blockItemId],
+  });
 
   console.log("Removed blockItem: ", blockItemText);
 }
@@ -261,11 +179,7 @@ async function removeBlockItems() {
 async function deleteBlocklist() {
   const blocklistName = "TestBlocklist";
 
-  const result = await client.path("/text/blocklists/{blocklistName}", blocklistName).delete();
-
-  if (isUnexpected(result)) {
-    throw result;
-  }
+  await blocklistClient.deleteTextBlocklist(blocklistName);
 
   console.log("Deleted blocklist: ", blocklistName);
 }