+++ sdk/contentsafety/ai-content-safety-rest/src/static-helpers/pagingHelpers.ts
@@ -1,7 +1,13 @@
 // Copyright (c) Microsoft Corporation.
 // Licensed under the MIT License.
 
-import { Client, createRestError, PathUncheckedResponse } from "@azure-rest/core-client";
+import {
+  Client,
+  createRestError,
+  OperationOptions,
+  operationOptionsToRequestParameters,
+  PathUncheckedResponse,
+} from "@azure-rest/core-client";
 import { RestError } from "@azure/core-rest-pipeline";
 
 /**
@@ -81,6 +87,7 @@ export interface BuildPagedAsyncIteratorOptions {
   nextLinkName?: string;
   nextLinkMethod?: "GET" | "POST";
   apiVersion?: string;
+  requestOptions?: OperationOptions;
 }
 
 /**
@@ -109,10 +116,11 @@ export function buildPagedAsyncIterator<
         result = await getInitialResponse();
       } else {
         const resolvedPageLink = apiVersion ? addApiVersionToUrl(pageLink, apiVersion) : pageLink;
+        const requestParameters = operationOptionsToRequestParameters(options.requestOptions ?? {});
         result =
           nextLinkMethod === "POST"
-            ? await client.pathUnchecked(resolvedPageLink).post()
-            : await client.pathUnchecked(resolvedPageLink).get();
+            ? await client.pathUnchecked(resolvedPageLink).post(requestParameters)
+            : await client.pathUnchecked(resolvedPageLink).get(requestParameters);
       }
       checkPagingRequest(result, expectedStatuses);
       const results = await processResponseBody(result as TResponse);